Player coordinates are now stored as doubles like everything.

+++ player.c
@@ -3,6 +3,6 @@
 #include "defs.h"
 #include "player.h"
 
-float player_x = PLAYER_START_X;
-float player_y = PLAYER_START_Y;
-float player_angle = PLAYER_START_ANGLE;
+double player_x = PLAYER_START_X;
+double player_y = PLAYER_START_Y;
+double player_angle = PLAYER_START_ANGLE;
